+++ src/handler.ts
@@ -1,4 +1,4 @@
-import type { Hook, Handler, Context, ComposedHandler } from './types'
+import type { Context } from './types'
 
 const json = new Headers()
 json.append('Content-Type', 'application/json')
@@ -7,95 +7,185 @@ export const jsonHeader = {
 }
 
 // We don't want to assign new variable to be used only once here
-export const mapResponse = (
+export const mapEarlyResponse = (
+	headerAvailable: boolean,
 	response: unknown,
-	request: Context,
+	context: Context,
 	status: number
 ) => {
-	switch (typeof response) {
-		case 'string':
-			return new Response(response, {
-				status,
-				headers: request.responseHeaders
-			})
-
-		case 'object':
-			if (response instanceof Error) return errorToResponse(response)
-			if (response instanceof Response) {
-				for (const [key, value] of request.responseHeaders.entries())
-					response.headers.append(key, value)
-
-				return response
-			}
-
-			request.responseHeaders.append('Content-Type', 'application/json')
-
-			return new Response(JSON.stringify(response), {
-				status,
-				headers: request.responseHeaders
-			})
-
-		// ? Maybe response or Blob
-		case 'function':
-			if (response instanceof Blob)
+	if (headerAvailable)
+		switch (typeof response) {
+			case 'string':
 				return new Response(response, {
 					status,
-					headers: request.responseHeaders
+					headers: context.responseHeaders
 				})
 
-			for (const [key, value] of request.responseHeaders.entries())
-				(response as unknown as Response).headers.append(key, value)
+			case 'object':
+				if (response instanceof Error)
+					return errorToResponse(response, context.responseHeaders)
+				if (response instanceof Response) {
+					for (const [
+						key,
+						value
+					] of context.responseHeaders.entries())
+						response.headers.append(key, value)
 
-			return response as unknown as Response
+					return response
+				}
 
-		case 'number':
-		case 'boolean':
-			return new Response(response.toString(), {
-				status,
-				headers: request.responseHeaders
-			})
+				context.responseHeaders.append(
+					'Content-Type',
+					'application/json'
+				)
 
-		default:
-			break
-	}
+				return new Response(JSON.stringify(response), {
+					status,
+					headers: context.responseHeaders
+				})
+
+			// ? Maybe response or Blob
+			case 'function':
+				if (response instanceof Blob)
+					return new Response(response, {
+						status,
+						headers: context.responseHeaders
+					})
+
+				for (const [key, value] of context.responseHeaders.entries())
+					(response as unknown as Response).headers.append(key, value)
+
+				return response as unknown as Response
+
+			case 'number':
+			case 'boolean':
+				return new Response(response.toString(), {
+					status,
+					headers: context.responseHeaders
+				})
+
+			default:
+				break
+		}
+	else
+		switch (typeof response) {
+			case 'string':
+				return new Response(response)
+
+			case 'object':
+				if (response instanceof Response) return response
+				if (response instanceof Error)
+					return errorToResponse(response, context.responseHeaders)
+
+				return new Response(JSON.stringify(response), jsonHeader)
+
+			// ? Maybe response or Blob
+			case 'function':
+				if (response instanceof Blob) return new Response(response)
+
+				return response as unknown as Response
+
+			case 'number':
+			case 'boolean':
+				return new Response(response.toString())
+
+			default:
+				break
+		}
 }
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export const mapResponseWithoutHeaders = (
+export const mapResponse = (
+	headerAvailable: boolean,
 	response: unknown,
-	request?: Context,
-	status?: number
+	context: Context,
+	status: number
 ) => {
-	switch (typeof response) {
-		case 'string':
-			return new Response(response)
+	if (headerAvailable)
+		switch (typeof response) {
+			case 'string':
+				return new Response(response, {
+					status,
+					headers: context.responseHeaders
+				})
 
-		case 'object':
-			if (response instanceof Response) return response
-			if (response instanceof Error) return errorToResponse(response)
+			case 'object':
+				if (response instanceof Error)
+					return errorToResponse(response, context.responseHeaders)
+				if (response instanceof Response) {
+					for (const [key, value] of context.responseHeaders!.entries())
+						response.headers.append(key, value)
 
-			return new Response(JSON.stringify(response), jsonHeader)
+					return response
+				}
 
-		// ? Maybe response or Blob
-		case 'function':
-			if (response instanceof Blob) return new Response(response)
+				context.responseHeaders.append(
+					'Content-Type',
+					'application/json'
+				)
 
-			return response as unknown as Response
+				return new Response(JSON.stringify(response), {
+					status,
+					headers: context.responseHeaders
+				})
 
-		case 'number':
-		case 'boolean':
-			return new Response(response.toString())
+			// ? Maybe response function or Blob
+			case 'function':
+				if (response instanceof Blob)
+					return new Response(response, {
+						status,
+						headers: context.responseHeaders
+					})
 
-		default:
-			break
-	}
+				return response()
+
+			case 'number':
+			case 'boolean':
+				return new Response(response.toString(), {
+					status,
+					headers: context.responseHeaders
+				})
+
+			case 'undefined':
+				return new Response('', {
+					status,
+					headers: context.responseHeaders
+				})
+
+			default:
+				return new Response(response as any, {
+					status,
+					headers: context.responseHeaders
+				})
+		}
+	else
+		switch (typeof response) {
+			case 'string':
+				return new Response(response)
+
+			case 'object':
+				if (response instanceof Response) return response
+				if (response instanceof Error)
+					return errorToResponse(response, context.responseHeaders)
+
+				return new Response(JSON.stringify(response), jsonHeader)
+
+			// ? Maybe response or Blob
+			case 'function':
+				if (response instanceof Blob) return new Response(response)
+
+				return response()
+
+			case 'number':
+			case 'boolean':
+				return new Response(response.toString())
+
+			case 'undefined':
+				return new Response('')
+
+			default:
+				return new Response(response as any)
+		}
 }
-
-// Currying to merge hook
-export const composeHandler = (
-	handler: Handler<any, any>,
-	hook: Hook<any>
-): ComposedHandler => [handler, hook]
 
 export const errorToResponse = (error: Error, headers?: HeadersInit) =>
 	new Response(